requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
Pillow==10.1.0
torch>=2.2.0
torchvision>=0.17.0
ultralytics>=8.0.196
numpy>=1.24.3
python-multipart==0.0.6
requests>=2.31.0 